moving movie.js around and minor fixes
